fixes to boss logic
