+++ src/helpers/createJoker.ts
@@ -4,6 +4,8 @@ export const createJoker = (
 	room: string,
 	event: string
 ): string[] => {
+	if (codes.length == 0)
+		return []
 	return [`patch=1,EE,E0${(codes.length + 3)
 			.toString(16)
 			.padStart(2, "0")
